@@ -1,15 +1,16 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+	  <TargetFramework>net6.0</TargetFramework>
+	  <Version>6.0.0</Version>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>bin\Debug\netcoreapp2.1\Ae.B.User.Api.xml</DocumentationFile>
+    <DocumentationFile>bin\Debug\net6.0\Ae.B.User.Api.xml</DocumentationFile>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <DocumentationFile>bin\Release\netcoreapp2.1\Ae.B.User.Api.xml</DocumentationFile>
+    <DocumentationFile>bin\Release\net6.0\Ae.B.User.Api.xml</DocumentationFile>
   </PropertyGroup>
 
   <ItemGroup>
@@ -17,14 +18,14 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="log4net" Version="2.0.8" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.10" />
-    <PackageReference Include="NLog" Version="4.6.7" />
-    <PackageReference Include="NLog.Web.AspNetCore" Version="4.9.0" />
-    <PackageReference Include="ApolloErp.Login.Auth" Version="1.0.0" />
-    <PackageReference Include="ApolloErp.Web" Version="1.0.1" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
+    <PackageReference Include="ApolloErp.Login.Auth" Version="6.0.0" />
+    <PackageReference Include="ApolloErp.Web" Version="6.0.0" />
+	  <PackageReference Include="log4net" Version="2.0.17" />
+	  <PackageReference Include="NLog" Version="5.3.2" />
+	  <PackageReference Include="NLog.Web.AspNetCore" Version="5.3.11" />
+	  <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="6.0.16" />
+	  <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.1" />
+	  <PackageReference Include="Microsoft.Extensions.DependencyModel" Version="8.0.0" />
   </ItemGroup>
 
   <ItemGroup>
